requirements.txt
selenium==4.16.0
requests
beautifulsoup4
python-dotenv
googletrans==4.0.0-rc1
webdriver-manager
browserstack-local
browserstack-sdk